refactor: Post, Skeleton 코드 분리
