+++ client/src/components/HomePage/index.tsx
@@ -1,9 +1,10 @@
 import PostWriter from 'components/HomePage/PostWriter';
 import PostWriterModal from './PostWriterModal';
-import Post from 'components/HomePage/Post';
+import Post from './Post';
 import PostImageBox from 'components/HomePage/PostImageBox';
 import PostList from 'components/HomePage/PostList';
 import ImageViewer from 'components/HomePage/ImageViewer';
+import Skeleton from './Skeleton';
 
 export {
   PostWriter,
@@ -11,5 +12,6 @@ export {
   Post,
   PostImageBox,
   PostList,
-  ImageViewer
+  ImageViewer,
+  Skeleton
 };
